more constants

+++ libs/openFrameworks/graphics/ofGraphics.cpp
@@ -397,8 +397,8 @@ void ofBackgroundGradient(const ofColor& start, const ofColor& end, ofGradientMo
 		glm::vec2 center(w / 2, h / 2);
 		gradientMesh.addVertex(glm::vec3(center, 0.f));
 		gradientMesh.addColor(start);
-		int n = 32; // circular gradient resolution
-		float angleBisector = glm::two_pi<float>() / (n * 2);
+		float n = 32; // circular gradient resolution
+		float angleBisector = glm::two_pi<float>() / (n * 2.0);
 		float smallRadius = ofDist(0, 0, w / 2, h / 2);
 		float bigRadius = smallRadius / cos(angleBisector);
 		for(int i = 0; i <= n; i++) {
